+++ examples/inverse/plot_dics_beamformer.py
@@ -3,9 +3,9 @@
 Compute DICS beamfomer on evoked data
 =====================================
 
-Compute a Dynamic Imaging of Coherent Sources (DICS) filter from single trial
-activity at frequencies between 8 and 12 Hz and estimate source time course on
-evoked data.
+Compute a Dynamic Imaging of Coherent Sources (DICS) beamformer from single
+trial activity in a time-frequency window to estimate source time courses based
+on evoked data.
 
 The original reference for DICS is:
 Gross et al. Dynamic imaging of coherent sources: Studying neural interactions
@@ -25,7 +25,7 @@ import numpy as np
 
 from mne.fiff import Raw
 from mne.datasets import sample
-from mne.time_frequency import compute_csd
+from mne.time_frequency import compute_epochs_csd
 from mne.beamformer import dics
 
 data_path = sample.data_path()
@@ -59,10 +59,10 @@ forward = mne.read_forward_solution(fname_fwd, surf_ori=True)
 # Computing the data and noise cross-spectral density matrices
 # The time-frequency window was chosen on the basis of spectrograms from
 # example time_frequency/plot_time_frequency.py
-data_csd = compute_csd(epochs, mode='multitaper', tmin=0.04, tmax=0.15, fmin=6,
-                       fmax=10)
-noise_csd = compute_csd(epochs, mode='multitaper', tmin=-0.11, tmax=0.0,
-                        fmin=6, fmax=10)
+data_csd = compute_epochs_csd(epochs, mode='multitaper', tmin=0.04, tmax=0.15,
+                              fmin=6, fmax=10)
+noise_csd = compute_epochs_csd(epochs, mode='multitaper', tmin=-0.11, tmax=0.0,
+                               fmin=6, fmax=10)
 
 evoked = epochs.average()
 
@@ -84,4 +84,6 @@ brain = stc.plot(surface='inflated', hemi='rh', subjects_dir=subjects_dir)
 brain.set_data_time_index(180)
 brain.scale_data_colormap(fmin=4, fmid=6, fmax=8, transparent=True)
 brain.show_view('lateral')
-brain.save_image('dSPM_map.png')
+
+# Uncomment to save image
+#brain.save_image('DICS_map.png')