+++ mne/beamformer/__init__.py
@@ -2,4 +2,4 @@
 """
 
 from ._lcmv import lcmv, lcmv_epochs, lcmv_raw
-from ._dics import dics, dics_epochs
+from ._dics import dics, dics_epochs, dics_source_power