+++ mne/beamformer/_dics.py
@@ -5,24 +5,25 @@
 #
 # License: BSD (3-clause)
 
+import warnings
+
 import numpy as np
 from scipy import linalg
 
 import logging
 logger = logging.getLogger('mne')
 
-from ..fiff.constants import FIFF
-from ..fiff.proj import make_projector
-from ..fiff.pick import pick_types, pick_channels_forward
+from ._lcmv import _prepare_beamformer_input
+from ..fiff.pick import pick_types
 from ..forward import _subject_from_forward
-from ..minimum_norm.inverse import _get_vertno, combine_xyz
+from ..minimum_norm.inverse import combine_xyz
 from ..source_estimate import SourceEstimate
-from ..source_space import label_src_vertno_sel
+from ..time_frequency import CrossSpectralDensity
 from .. import verbose
 
 
 @verbose
-def _apply_dics(data, info, tmin, forward, noise_csd, data_csd, reg=0.1,
+def _apply_dics(data, info, tmin, forward, noise_csd, data_csd, reg,
                 label=None, picks=None, pick_ori=None, verbose=None):
     """Dynamic Imaging of Coherent Sources (DICS).
 
@@ -41,9 +42,9 @@ def _apply_dics(data, info, tmin, forward, noise_csd, data_csd, reg=0.1,
         Time of first sample.
     forward : dict
         Forward operator.
-    noise_csd : CrossSpectralDensity
+    noise_csd : instance of CrossSpectralDensity
         The noise cross-spectral density.
-    data_csd : CrossSpectralDensity
+    data_csd : instance of CrossSpectralDensity
         The data cross-spectral density.
     reg : float
         The regularization for the cross-spectral density.
@@ -63,60 +64,14 @@ def _apply_dics(data, info, tmin, forward, noise_csd, data_csd, reg=0.1,
     stc : SourceEstimate (or list of SourceEstimate)
         Source time courses.
     """
-    # TODO: DICS, in the original 2001 paper, used a free orientation
-    # beamformer, however selection of the max-power orientation was also
-    # employed depending on whether a dominant component was present
 
-    is_free_ori = forward['source_ori'] == FIFF.FIFFV_MNE_FREE_ORI
+    is_free_ori, picks, _, proj, vertno, G =\
+        _prepare_beamformer_input(info, forward, label, picks, pick_ori)
 
-    # DIFF: LCMV has 'max-power' here in addition to 'normal'
-    if pick_ori == 'normal' and not is_free_ori:
-        raise ValueError('Normal orientation can only be picked when a '
-                         'forward operator with free orientation is used.')
-    if pick_ori == 'normal' and not forward['surf_ori']:
-        raise ValueError('Normal orientation can only be picked when a '
-                         'forward operator oriented in surface coordinates is '
-                         'used.')
-    if pick_ori == 'normal' and not forward['src'][0]['type'] == 'surf':
-        raise ValueError('Normal orientation can only be picked when a '
-                         'forward operator with a surface-based source space '
-                         'is used.')
-
-    if picks is None:
-        picks = pick_types(info, meg=True, eeg=True, exclude='bads')
-
-    ch_names = [info['ch_names'][k] for k in picks]
-
-    # Restrict forward solution to selected channels
-    forward = pick_channels_forward(forward, include=ch_names)
-
-    # Get gain matrix (forward operator)
-    if label is not None:
-        vertno, src_sel = label_src_vertno_sel(label, forward['src'])
-
-        if is_free_ori:
-            src_sel = 3 * src_sel
-            src_sel = np.c_[src_sel, src_sel + 1, src_sel + 2]
-            src_sel = src_sel.ravel()
-
-        G = forward['sol']['data'][:, src_sel]
-    else:
-        vertno = _get_vertno(forward['src'])
-        G = forward['sol']['data']
-
-    # Apply SSPs
-    proj = None
-    if 'projs' in info and len(info['projs']) > 0:
-        proj, ncomp, _ = make_projector(info['projs'], ch_names)
-        G = np.dot(proj, G)
-
-    # DIFF: LCMV applies SSPs and whitener to data covariance at this point,
-    # here we only read in the cross-spectral density matrix - the data used in
-    # its calculation already had SSPs applied and we will use the noise CSD
-    # matrix in noise normalization instead of whitening the data
     Cm = data_csd.data
 
-    # Cm += reg * np.trace(Cm) / len(Cm) * np.eye(len(Cm))
+    # Calculating regularized inverse, equivalent to an inverse operation after
+    # regularization: Cm += reg * np.trace(Cm) / len(Cm) * np.eye(len(Cm))
     Cm_inv = linalg.pinv(Cm, reg)
 
     # Compute spatial filters
@@ -129,8 +84,7 @@ def _apply_dics(data, info, tmin, forward, noise_csd, data_csd, reg=0.1,
         Gk = G[:, n_orient * k: n_orient * k + n_orient]
         Ck = np.dot(Wk, Gk)
 
-        # DIFF: LCMV calculates 'max-power' orientation here
-        # TODO: max-power is not used in this example, however DICS does employ
+        # TODO: max-power is not implemented yet, however DICS does employ
         # orientation picking when one eigen value is much larger than the
         # other
 
@@ -142,23 +96,14 @@ def _apply_dics(data, info, tmin, forward, noise_csd, data_csd, reg=0.1,
             Wk /= Ck
 
         # Noise normalization
-        # DIFF: LCMV prepares noise normalization outside of the loop
-        # DIFF: noise_norm is not complex in LCMV
         noise_norm = np.dot(np.dot(Wk.conj(), noise_csd.data), Wk.T)
         noise_norm = np.abs(noise_norm).trace()
         Wk /= np.sqrt(noise_norm)
-
-    # DIFF: LCMV picks 'max-power' orientation here
-
-    # DIFF: LCMV prepares noise normalization here and it doesn't involve the
-    # noise covariance
 
     # Pick source orientation normal to cortical surface
     if pick_ori == 'normal':
         W = W[2::3]
         is_free_ori = False
-
-    # DIFF: LCMV applies noise normalization for fixed orientation here
 
     if isinstance(data, np.ndarray) and data.ndim == 2:
         data = [data]
@@ -174,18 +119,15 @@ def _apply_dics(data, info, tmin, forward, noise_csd, data_csd, reg=0.1,
         if not return_single:
             logger.info("Processing epoch : %d" % (i + 1))
 
-        # DIFF: LCMV applies data whitening here
         # Apply SSPs
-        if proj is not None:
+        if info['projs']:
             M = np.dot(proj, M)
 
         # project to source space using beamformer weights
-
         if is_free_ori:
             sol = np.dot(W, M)
             logger.info('combining the current components...')
             sol = combine_xyz(sol)
-            # DIFF: LCMV applies noise normalization for free orientation here
         else:
             # Linear inverse: do not delay compuation due to non-linear abs
             sol = np.dot(W, M)
@@ -220,9 +162,9 @@ def dics(evoked, forward, noise_csd, data_csd, reg=0.01, label=None,
         Evoked data.
     forward : dict
         Forward operator.
-    noise_csd : CrossSpectralDensity
+    noise_csd : instance of CrossSpectralDensity
         The noise cross-spectral density.
-    data_csd : CrossSpectralDensity
+    data_csd : instance of CrossSpectralDensity
         The data cross-spectral density.
     reg : float
         The regularization for the cross-spectral density.
@@ -275,9 +217,9 @@ def dics_epochs(epochs, forward, noise_csd, data_csd, reg=0.01, label=None,
         Single trial epochs.
     forward : dict
         Forward operator.
-    noise_csd : CrossSpectralDensity
+    noise_csd : instance of CrossSpectralDensity
         The noise cross-spectral density.
-    data_csd : CrossSpectralDensity
+    data_csd : instance of CrossSpectralDensity
         The data cross-spectral density.
     reg : float
         The regularization for the cross-spectral density.
@@ -318,3 +260,145 @@ def dics_epochs(epochs, forward, noise_csd, data_csd, reg=0.01, label=None,
         stcs = list(stcs)
 
     return stcs
+
+
+@verbose
+def dics_source_power(info, forward, noise_csds, data_csds, reg=0.01,
+                      label=None, picks=None, pick_ori=None, verbose=None):
+    """Dynamic Imaging of Coherent Sources (DICS).
+
+    Calculate source power in time and frequency windows specified in the
+    calculation of the data cross-spectral density matrix or matrices. Source
+    power is normalized by noise power.
+
+    NOTE : This implementation has not been heavilly tested so please
+    report any issues or suggestions.
+
+    Parameters
+    ----------
+    info : dict
+        Measurement info, e.g. epochs.info.
+    forward : dict
+        Forward operator.
+    noise_csds : instance or list of instances of CrossSpectralDensity
+        The noise cross-spectral density matrix for a single frequency or a
+        list of matrices for multiple frequencies.
+    data_csds : instance or list of instances of CrossSpectralDensity
+        The data cross-spectral density matrix for a single frequency or a list
+        of matrices for multiple frequencies.
+    reg : float
+        The regularization for the cross-spectral density.
+    label : Label | None
+        Restricts the solution to a given label.
+    pick_ori : None | 'normal'
+        If 'normal', rather than pooling the orientations by taking the norm,
+        only the radial component is kept.
+    verbose : bool, str, int, or None
+        If not None, override default verbose level (see mne.verbose).
+
+    Returns
+    -------
+    stc: SourceEstimate
+        Source power with frequency instead of time.
+
+    Notes
+    -----
+    The original reference is:
+    Gross et al. Dynamic imaging of coherent sources: Studying neural
+    interactions in the human brain. PNAS (2001) vol. 98 (2) pp. 694-699
+    """
+
+    if isinstance(data_csds, CrossSpectralDensity):
+        data_csds = [data_csds]
+
+    if isinstance(noise_csds, CrossSpectralDensity):
+        noise_csds = [noise_csds]
+
+    csd_shapes = lambda x: tuple(c.data.shape for c in x)
+    if (csd_shapes(data_csds) != csd_shapes(noise_csds) or
+       any([len(set(csd_shapes(c))) > 1 for c in [data_csds, noise_csds]])):
+        raise ValueError('One noise CSD matrix should be provided for each '
+                         'data CSD matrix and vice versa. All CSD matrices '
+                         'should have identical shape.')
+
+    frequencies = []
+    for data_csd, noise_csd in zip(data_csds, noise_csds):
+        if data_csd.frequencies != noise_csd.frequencies:
+            raise ValueError('Data and noise CSDs should be calculated at '
+                             'identical frequencies')
+
+        # If CSD is summed over multiple frequencies, take the average
+        # frequency
+        if(len(data_csd.frequencies) > 1):
+            frequencies.append(np.mean(data_csd.frequencies))
+        else:
+            frequencies.append(data_csd.frequencies[0])
+    fmin = frequencies[0]
+
+    if len(frequencies) > 2:
+        fstep = []
+        for i in range(len(frequencies) - 1):
+            fstep.append(frequencies[i+1] - frequencies[i])
+        if not np.allclose(fstep, np.mean(fstep), 1e-5):
+            warnings.warn('Uneven frequency spacing in CSD object, '
+                          'frequencies in the resulting stc file will be '
+                          'inaccurate.')
+        fstep = fstep[0]
+    elif len(frequencies) > 1:
+        fstep = frequencies[1] - frequencies[0]
+    else:
+        fstep = 1  # dummy value
+
+    is_free_ori, picks, _, proj, vertno, G =\
+        _prepare_beamformer_input(info, forward, label, picks=None,
+                                  pick_ori=pick_ori)
+
+    n_orient = 3 if is_free_ori else 1
+    n_sources = G.shape[1] // n_orient
+    source_power = np.zeros((n_sources, len(data_csds)))
+    n_csds = len(data_csds)
+
+    logger.info('Computing DICS source power...')
+    for i, (data_csd, noise_csd) in enumerate(zip(data_csds, noise_csds)):
+        if n_csds > 1:
+            logger.info('    computing DICS spatial filter %d out of %d' %
+                        (i + 1, n_csds))
+
+        Cm = data_csd.data
+
+        # Calculating regularized inverse, equivalent to an inverse operation
+        # after the following regularization:
+        # Cm += reg * np.trace(Cm) / len(Cm) * np.eye(len(Cm))
+        Cm_inv = linalg.pinv(Cm, reg)
+
+        # Compute spatial filters
+        W = np.dot(G.T, Cm_inv)
+        for k in xrange(n_sources):
+            Wk = W[n_orient * k: n_orient * k + n_orient]
+            Gk = G[:, n_orient * k: n_orient * k + n_orient]
+            Ck = np.dot(Wk, Gk)
+
+            if is_free_ori:
+                # Free source orientation
+                Wk[:] = np.dot(linalg.pinv(Ck, 0.1), Wk)
+            else:
+                # Fixed source orientation
+                Wk /= Ck
+
+            # Noise normalization
+            noise_norm = np.dot(np.dot(Wk.conj(), noise_csd.data), Wk.T)
+            noise_norm = np.abs(noise_norm).trace()
+
+            # Calculating source power
+            sp_temp = np.dot(np.dot(Wk.conj(), data_csd.data), Wk.T)
+            sp_temp /= noise_norm
+            if pick_ori == 'normal':
+                source_power[k, i] = np.abs(sp_temp)[2, 2]
+            else:
+                source_power[k, i] = np.abs(sp_temp).trace()
+
+    logger.info('[done]')
+
+    subject = _subject_from_forward(forward)
+    return SourceEstimate(source_power, vertices=vertno, tmin=fmin / 1000.,
+                          tstep=fstep / 1000., subject=subject)